make file transfer concurrent, cleanup unused code, add readme
Using generators for file transfer instead of default method, makes it
possible to just use 'threading=True' in server configuration. While
not recommended for production, this method seems good enough for my
needs.

On the fly convertion and old file contents fetching are long abandoned
and were removed. Also some queue modifications were not needed in the
end and it was simpler to reuse super() implementation.

Readme draft for github.